UPSTREAM: mwifiex: fix misnomers in mwifiex_free_lock_list()
Despite the name (and meticulous comments), this function frees no
memory and does not touch any locks. All it does is "delete" the list
heads -- which just means they'll be dangling, and we'll need to re-init
them if we use them again.

It seems like this code would work OK as a sort of canary for using the
list after we've torn everything down, so it's fine to keep the code;
let's just get the name and comments to match what's actually happening.

BUG=b:37946775
TEST=kevin Wifi with KASAN enabled; reset tests + matfunc

Change-Id: I723c1f1102bca7f1f04a67803e56e63e7661d8a9
Signed-off-by: Brian Norris <briannorris@chromium.org>
Signed-off-by: Kalle Valo <kvalo@codeaurora.org>
Signed-off-by: Guenter Roeck <groeck@chromium.org>
(cherry picked from commit 5e6588b9d4ab)