UPSTREAM: HID: input: optionally use device id in battery name
Manufacturers do not always populate serial number in their devices, so
let's fall back to device ID when forming the battery device name. As a
result, batteries in devices without serial number will be named like
this:

	hid-0018:2D1F:510E.0001-battery

(as opposed to hid--battery for the first one, and failing to create
batteries for the subsequent ones).

BUG=chromium:733740,b:62674499
TEST=Build and boot eve

Change-Id: Icf2771a79711d290d20c66fdf635050bbff97843
Signed-off-by: Dmitry Torokhov <dmitry.torokhov@gmail.com>
Signed-off-by: Jiri Kosina <jkosina@suse.cz>
Signed-off-by: Guenter Roeck <groeck@chromium.org>
(cherry picked from commit 0922386538f9)