FROMLIST: HID: input: optionally use device id in battery name
Manufacturers do not always populate serial number in their devices, so
let's fall back to device ID when forming the battery device name. As a
result, batteries in devices without serial number will be named like
this:

	hid-0018:2D1F:510E.0001-battery

(as opposed to hid--battery for the first one, and failing to create
batteries for the subsequent ones).

Signed-off-by: Dmitry Torokhov <dmitry.torokhov@gmail.com>

BUG=chromium:733740,b:62674499
TEST=Build and boot eve

(am from https://patchwork.kernel.org/patch/9875635/)
Signed-off-by: Dmitry Torokhov <dtor@chromium.org>
Change-Id: Icf2771a79711d290d20c66fdf635050bbff97843
Reviewed-on: https://chromium-review.googlesource.com/596812
Reviewed-by: Benson Leung <bleung@chromium.org>