net: danipc: Use polling on netdev interface
To improve performance, the network device interface will be polled
instead of being interrupt driven. This eliminates being interrupted
on every packet posted to the hwfifo in the UL.

Significant code clean-up is done to provide the flexibility of HW
FIFO mapping and simplify the implementation on both network and
character device interface.

CRs-Fixed: 2072979
Change-Id: I0fc23ae599aec3b598da7516b7cd2e90346a072a
Acked-by: Jie Luo <jluo@qti.qualcomm.com>
Signed-off-by: Yasir Malik <ymalik@codeaurora.org>