FROMGIT: bitops: avoid integer overflow in GENMASK(_ULL)
GENMASK(_ULL) performs a left-shift of ~0UL(L), which technically results
in an integer overflow.  clang raises a warning if the overflow occurs in
a preprocessor expression.  Clear the low-order bits through a
substraction instead of the left-shift to avoid the overflow.

(akpm: no change in .text size in my testing)

(cherry picked from git.kernel.org next/linux-next akpm
commit 041800ec827a152938595aaa038213dfe6ecb025)

Change-Id: I002b511cb2a80da6d9a7734c58de15ed9d5bb526
Link: http://lkml.kernel.org/r/20170803212020.24939-1-mka@chromium.org
Signed-off-by: Matthias Kaehlcke <mka@chromium.org>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>
Signed-off-by: Stephen Rothwell <sfr@canb.auug.org.au>