fs/select.c: fix memory corruption in compat_get_fd_set()
464d62421cb8 ("select: switch compat_{get,put}_fd_set() to
compat_{get,put}_bitmap()") changed the calculation on how many bytes need
to be zeroed when userspace handed over a NULL pointer for a fdset array
in the select syscall.

The calculation was changed in compat_get_fd_set() wrongly from
	memset(fdset, 0, ((nr + 1) & ~1)*sizeof(compat_ulong_t));
to
	memset(fdset, 0, ALIGN(nr, BITS_PER_LONG));

The ALIGN(nr, BITS_PER_LONG) calculates the number of bits which need to
be zeroed in the target fdset array (rounded up to the next full bits for
an unsigned long).  But the memset() call expects the number of bytes to
be zeroed.

This leads to clearing more memory than wanted (on the stack area or even
at kmalloc()ed memory areas) and to random kernel crashes as we have seen
them on the parisc platform.

The correct change should have been
	memset(fdset, 0, (ALIGN(nr, BITS_PER_LONG) / BITS_PER_LONG) * BYTES_PER_LONG);
which is the same as can be archieved with a call to
	zero_fd_set(nr, fdset).

Link: http://lkml.kernel.org/r/20170823203700.GA29614@ls3530.fritz.box
Fixes: 464d62421cb8 ("select: switch compat_{get,put}_fd_set() to compat_{get,put}_bitmap()"
Signed-off-by: Helge Deller <deller@gmx.de>
Cc: Al Viro <viro@zeniv.linux.org.uk>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>
Signed-off-by: Stephen Rothwell <sfr@canb.auug.org.au>