btrfs: Remove received_uuid during received snapshot ro->rw switch
Currently when a read-only snapshot is received and subsequently its ro
property is set to false i.e. switched to rw-mode the received_uuid of
that subvol remains intact. However, once the received volume is
switched to RW mode we cannot guaranteee that it contains the same data,
so it makes sense to remove the received uuid. The presence of the
received_uuid can also cause problems when the volume is being sent.

Signed-off-by: Nikolay Borisov <nborisov@suse.com>
Suggested-by: David Sterba <dsterba@suse.cz>
Signed-off-by: David Sterba <dsterba@suse.com>