linux-next: drm/i916: fix nop_submit_request mismerge
A mismerge in linux-next caused a build failure:

drivers/gpu/drm/i915/i915_gem.c: In function 'nop_submit_request':
drivers/gpu/drm/i915/i915_gem.c:3092:54: error: 'flags' undeclared (first use in this function); did you mean 'class'?

This is the fixup I used to get a clean build on top of today's
tree. I don't know which of the versions is the one we want,
so this may just as well be wrong.

Fixes: 05724bd054c3 ("Merge branch 'akpm/master'")
Signed-off-by: Arnd Bergmann <arnd@arndb.de>
Signed-off-by: Mark Brown <broonie@kernel.org>