locking/barriers: Convert users of lockless_dereference() to READ_ONCE()
This is a limited cherry-pick of the following commit, restricted to arch/x86/include/asm/mmu_context.h:

  506458efaf15 ("locking/barriers: Convert users of lockless_dereference() to READ_ONCE()")

This cherry-pick is correct for x86 even without the other dependent commit,
and simplifies the conflict-free merging of the Page Table Isolation feature.

READ_ONCE() now has an implicit smp_read_barrier_depends() call, so it
can be used instead of lockless_dereference() without any change in
semantics.

Signed-off-by: Will Deacon <will.deacon@arm.com>
Cc: Linus Torvalds <torvalds@linux-foundation.org>
Cc: Paul E. McKenney <paulmck@linux.vnet.ibm.com>
Cc: Peter Zijlstra <peterz@infradead.org>
Cc: Thomas Gleixner <tglx@linutronix.de>
Link: http://lkml.kernel.org/r/1508840570-22169-4-git-send-email-will.deacon@arm.com
Signed-off-by: Ingo Molnar <mingo@kernel.org>