ARM: BUG if jumping to usermode address in kernel mode
Detect if we are returning to usermode via the normal kernel exit paths
but the saved PSR value indicates that we are in kernel mode.  This
could occur due to corrupted stack state, which has been observed with
"ftracetest".

This ensures that we catch the problem case before we get to user code.

Signed-off-by: Russell King <rmk+kernel@armlinux.org.uk>
(cherry picked from commit 8bafae202c82dc257f649ea3c275a0f35ee15113)
Signed-off-by: Alex Shi <alex.shi@linaro.org>