eeprom: at24: don't use the default regmap locking
We need to always wait for the last write to complete and we want the
regmap to be protected from concurrent access during that wait. That
makes using the regmap's default locking impossible and we're
protecting it with an external mutex.

Set an empty function as the callback for regmap's lock and unlock
routines so that it doesn't allocate any internal resources for
unnecessary locking.

Signed-off-by: Bartosz Golaszewski <brgl@bgdev.pl>