arch/tile: mark as orphaned
The chip family of TILEPro and TILE-Gx was developed by Tilera,
which was eventually acquired by Mellanox.  The tile architecture
was added to the kernel in 2010 and first appeared in 2.6.36.
Now at Mellanox we are developing new chips based on the ARM64
architecture; our last TILE-Gx chip (the Gx72) was released in 2013,
and our customers using tile architecture products are not, as far
as we know, looking to upgrade to newer kernel releases.  In the
absence of someone in the community stepping up to take over
maintainership, this commit marks the architecture as orphaned.

Cc: Chris Metcalf <metcalf@alum.mit.edu>
Signed-off-by: Chris Metcalf <cmetcalf@mellanox.com>