x86/entry: Remove SYSENTER_stack naming
If the kernel oopses while on the trampoline stack, it will say
"<SYSENTER>" even if SYSENTER is not involved.  That is rather
confusing.

The "SYSENTER" stack is used for a lot more than SYSENTER now.
Give it a better string to display in stack dumps, and rename
the kernel code to match.

We also move the 32-bit code over to the new naming even though
it still uses the entry stack only for SYSENTER.

Signed-off-by: Dave Hansen <dave.hansen@linux.intel.com>
Cc: Andy Lutomirski <luto@kernel.org>
Cc: Borislav Petkov <bp@alien8.de>
Cc: Borislav Petkov <bp@suse.de>
Cc: Brian Gerst <brgerst@gmail.com>
Cc: Denys Vlasenko <dvlasenk@redhat.com>
Cc: H. Peter Anvin <hpa@zytor.com>
Cc: Josh Poimboeuf <jpoimboe@redhat.com>
Cc: Linus Torvalds <torvalds@linux-foundation.org>
Cc: Peter Zijlstra <peterz@infradead.org>
Cc: Rik van Riel <riel@redhat.com>
Cc: Thomas Gleixner <tglx@linutronix.de>
Link: http://lkml.kernel.org/r/20171205012507.3F55C4DD@viggo.jf.intel.com
Signed-off-by: Ingo Molnar <mingo@kernel.org>