mm/vmscan.c: terminate shrink_slab loop if signal is pending
Slab shrinkers can be quite time consuming and when signal is pending they
can delay handling of the signal.  If fatal signal is pending there is no
point in shrinking that process since it will be killed anyway.  This
change checks for pending fatal signals inside shrink_slab loop and if one
is detected terminates this loop early.

The observable delay depends on the implementation of the shrinkers
registered in the system including the ones from drivers.  I've
captured traces showing delays of up to 100ms where the process with
pending SIGKILL is in direct memory reclaim and signal handling is
delayed because of that.  I realize that it's not the fault of
shrink_slab() that some shrinkers take long time to shrink their slabs
(sometimes because of justifiable reasons and sometimes because of a
bug which has to be fixed) but this can be a safeguard against such
cases.  Couple shrinker examples that I found most time consuming are
(most of that 100ms delay is the result of the first two ones):

https://patchwork.kernel.org/patch/10096641/: the patch fixes dm-bufio
shrinker which in certain conditions reclaims only one buffer per scan
making the shrinking process very inefficient.

https://android.googlesource.com/kernel/msm/+/android-7.1.0_r0.2/drivers/gpu/msm/kgsl_pool.c#420:
This example is from a driver where shrinker returns 0 instead of
SHRINK_STOP when it's unable to reclaim anymore.  As a result when
total_scan in do_shrink_slab() is large this will cause multiple
scan_objects() calls with no memory being reclaimed.  Patch for this
one is under review by the owners.

Shrinker that seems to be justifiably heavy is super_cache_scan()
inside fs/super.c.  I have traces where it takes up to 4ms to complete.

Link: http://lkml.kernel.org/r/20171206192026.25133-1-surenb@google.com
Signed-off-by: Suren Baghdasaryan <surenb@google.com>
Cc: Michal Hocko <mhocko@suse.com>
Cc: Johannes Weiner <hannes@cmpxchg.org>
Cc: Hillf Danton <hdanton@sina.com>
Cc: Minchan Kim <minchan@kernel.org>
Cc: Mel Gorman <mgorman@techsingularity.net>
Cc: Huang Ying <ying.huang@intel.com>
Cc: Tim Murray <timmurray@google.com>
Cc: Todd Kjos <tkjos@google.com>
Cc: Christoph Lameter <cl@linux.com>
Cc: Pekka Enberg <penberg@kernel.org>
Cc: David Rientjes <rientjes@google.com>
Cc: Joonsoo Kim <iamjoonsoo.kim@lge.com>
Cc: Sergey Senozhatsky <sergey.senozhatsky.work@gmail.com>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>
Signed-off-by: Stephen Rothwell <sfr@canb.auug.org.au>