mm, hugetlbfs: introduce ->pagesize() to vm_operations_struct
When device-dax is operating in huge-page mode we want it to behave like
hugetlbfs and report the MMU page mapping size that is being enforced by
the vma.  Similar to 31383c6865a5 ("mm, hugetlbfs: introduce ->split() to
vm_operations_struct") it would be messy to teach vma_mmu_pagesize() about
device-dax page mapping sizes in the same (hstate) way that hugetlbfs
communicates this attribute.  Instead, these patches introduce a new
->pagesize() vm operation.

Link: http://lkml.kernel.org/r/151270385525.21215.16828596212056611775.stgit@dwillia2-desk3.amr.corp.intel.com
Signed-off-by: Dan Williams <dan.j.williams@intel.com>
Reported-by: Jane Chu <jane.chu@oracle.com>
Cc: Benjamin Herrenschmidt <benh@kernel.crashing.org>
Cc: Paul Mackerras <paulus@samba.org>
Cc: Michael Ellerman <mpe@ellerman.id.au>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>
Signed-off-by: Stephen Rothwell <sfr@canb.auug.org.au>